исправила на int feature-BJS2-96661-add-swagger-fix

+++ src/main/java/school/faang/user_service/service/user/UserSubscriptionServiceImpl.java
@@ -60,7 +60,7 @@ public class UserSubscriptionServiceImpl implements UserSubscriptionService {
 
     @Override
     public CountResponse getFollowersCount(long followeeId) {
-        long count = subscriptionRepository.findFollowersAmountByFolloweeId(followeeId);
+        int count = subscriptionRepository.findFollowersAmountByFolloweeId(followeeId);
         log.debug("Number of subscribers of the user {}: {}", followeeId, count);
         return new CountResponse(count);
     }
